+++ tests/test_dbus.py
@@ -151,25 +151,13 @@ def test__schedule_shutdown_fail(caplog, action, exception_message, log_message)
     assert "inhibitor" in caplog.records[2].message
 
 
-@pytest.mark.parametrize(
-    ("topic_suffix", "expected_action_arg"), [("poweroff", "poweroff")]
-)
-def test_mqtt_topic_suffix_action_mapping(topic_suffix, expected_action_arg):
-    mqtt_action = systemctl_mqtt._MQTT_TOPIC_SUFFIX_ACTION_MAPPING[topic_suffix]
+def test_lock_all_sessions(caplog):
     login_manager_mock = unittest.mock.MagicMock()
     with unittest.mock.patch(
-        "systemctl_mqtt._dbus.get_login_manager", return_value=login_manager_mock,
-    ):
-        mqtt_action.trigger(
-            state=systemctl_mqtt._State(
-                mqtt_topic_prefix="systemctl/hostname",
-                homeassistant_discovery_prefix="homeassistant",
-                homeassistant_node_id="node",
-                poweroff_delay=datetime.timedelta(),
-            )
-        )
-    assert login_manager_mock.ScheduleShutdown.call_count == 1
-    schedule_args, schedule_kwargs = login_manager_mock.ScheduleShutdown.call_args
-    assert len(schedule_args) == 2
-    assert schedule_args[0] == expected_action_arg
-    assert not schedule_kwargs
+        "systemctl_mqtt._dbus.get_login_manager", return_value=login_manager_mock
+    ), caplog.at_level(logging.INFO):
+        systemctl_mqtt._dbus.lock_all_sessions()
+    login_manager_mock.LockSessions.assert_called_once_with()
+    assert len(caplog.records) == 1
+    assert caplog.records[0].levelno == logging.INFO
+    assert caplog.records[0].message == "instruct all sessions to activate screen locks"